pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "venice-kb-collector"
version = "0.1.0"
description = "Pull, collate, and track changes in Venice AI API documentation for agent knowledge bases"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Venice KB Collector Contributors"}
]
readme = "README.md"
keywords = ["venice", "documentation", "knowledge-base", "api-docs", "changelog"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "beautifulsoup4>=4.12",
    "playwright>=1.45",
    "openai>=1.30",
    "tiktoken>=0.7",
    "pydantic>=2.7",
    "typer[all]>=0.12",
    "rich>=13.7",
    "python-dotenv>=1.0",
    "markdownify>=0.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "black>=24.0",
    "mypy>=1.10",
]

[project.scripts]
venice-kb = "venice_kb.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
